// frontend/app/services/analytics/types.gen.ts
export type DailyAggregationItem = {
    /**
     * Occupied Count
     */
    occupied_count: number;
    /**
     * Day
     */
    day: string;
};
// ...
export type HourlyAggregationItem = {
    /**
     * Occupied Count
     */
    occupied_count: number;
    /**
     * Hour
     */
    hour: string;
};